Add checksumming of HTML data
Previously this script would be run manually every month on the first to regenerate the podcast feed. This was silly.

The script now stores the checksum of the evening talks HTML document, and compares the checksum of the retrieved version to the stored checksum. The podcast feed XML will be regenerated only if the checksums differ. This is in preparation for setting this script to run autonomously.